refactor: split my-page service boundary
